requirements.txt
django
djangorestframework
corsheaders
gunicorn
